feat: Refactor listing data handling; introduce Listing interface, streamline addListing and updateListing functions, and enhance validation schema for improved data integrity
