+++ server/listing/listing.router.ts
@@ -1,7 +1,6 @@
 import express from "express";
 import type { Request, Response, NextFunction } from "express";
 const router = express.Router();
-import Joi from "joi";
 import bodyValidationMiddleware from "../middleware/bodyValidationMiddleware";
 import {
 	addListing,
@@ -11,6 +10,7 @@ import {
 } from "./listing.repository";
 
 import type { ListingDetails } from "./listing.repository";
+import { listingSchema } from "./listing.schema";
 
 router.use(express.json());
 
@@ -59,45 +59,6 @@ router.get(
 	},
 );
 
-// Schema for adding a new listing
-const listingSchema = Joi.object().keys({
-	listing: Joi.object({
-		titleCategory: Joi.object({
-			id: Joi.number().optional(),
-			title: Joi.string().required(),
-			subTitle: Joi.string(),
-			categoryId: Joi.number().greater(0).required(),
-			subCategoryId: Joi.number(),
-			endDate: Joi.date().greater("now").required(),
-		}).required(),
-		itemDetails: Joi.object({
-			description: Joi.string().required(),
-			condition: Joi.boolean().required(),
-		}).required(),
-		pricePayment: Joi.object({
-			listingPrice: Joi.string().required(),
-			reservePrice: Joi.string(),
-			creditCardPayment: Joi.boolean().required(),
-			bankTransferPayment: Joi.boolean().required(),
-			bitcoinPayment: Joi.boolean().required(),
-		}).required(),
-		shipping: Joi.object({
-			pickUp: Joi.boolean().required(),
-			shippingOption: Joi.string().required(),
-		}).required(),
-	})
-		.required()
-		.custom((value) => {
-			if (
-				value.pricePayment.creditCardPayment === false &&
-				value.pricePayment.bankTransferPayment === false &&
-				value.pricePayment.bitcoinPayment === false
-			) {
-				throw new Error("At least one of the payment methods must be selected");
-			}
-		}),
-});
-
 /**
  * Add a new Listing
  *
@@ -112,10 +73,9 @@ router.post(
 	bodyValidationMiddleware(listingSchema),
 	async (req, res, next) => {
 		try {
-			const listing = req.body.listing;
-
+			// Changed: use req.body directly
+			const listing = req.body;
 			const addListingResponse = await addListing(listing);
-
 			return res.json(addListingResponse);
 		} catch (err) {
 			console.error(err);
