+++ server/middleware/errorHandlerMiddleware.ts
@@ -11,20 +11,19 @@ interface Error {
 	message: string;
 }
 
-interface Request {
-	// Define properties of the request object as needed
-}
-
 interface Response {
 	status: (code: number) => Response;
-	json: (body: any) => void;
+	json: (body: unknown) => void;
 }
 
-interface NextFunction {
-	(): void;
-}
+type NextFunction = () => void;
 
-export default function errorHandlerMiddleware(err: Error, req: Request, res: Response, next: NextFunction) {
+export default function errorHandlerMiddleware(
+	err: Error,
+	req: Request,
+	res: Response,
+	next: NextFunction,
+) {
 	res.status(err.status || 500).json({
 		message: err.message,
 	});
